test(studio): isolate data selection setters
